+++ tests/cmd.generate.test.js
@@ -13,17 +13,8 @@ const {
 
 beforeEach(async () => {
   await setupBeforeEach();
-  // let result;
-  // result = await snapdev(`logout --force`);
-  // result = await snapdev(`login --username ${username} --password ${password}`);
-  // result = await snapdev(`deregister --force`);
 });
-afterEach(async () => {
-  // let result;
-  // result = await snapdev(`logout --force`);
-  // result = await snapdev(`login --username ${username} --password ${password}`);
-  // result = await snapdev(`deregister --force`);
-});
+afterEach(async () => {});
 
 test('snapdev generate', async () => {
   let result;
